api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
chromadb==0.5.23
sentence-transformers==3.3.1
rank-bm25==0.2.2
pydantic==2.10.3
anthropic==0.42.0
sse-starlette==2.1.3
python-dotenv==1.0.1
